requirements.txt
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
html5lib==1.0.1
idna==2.8
pkg-resources==0.0.0
python-dotenv==0.14.0
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
urllib3==1.25.3
webencodings==0.5.1
